fix: widen ServiceOAuthNonce primary key to bigint
The nonce store is high-churn (one row per signed OAuth request), so a 32-bit
AUTO_INCREMENT id could eventually exhaust the int range even though rows are
pruned (the counter is not reset by deletes). Use bigint for the id column and
map it as Long on the entity.

+++ src/main/java/io/github/carlos_emr/carlos/commn/model/ServiceOAuthNonce.java
@@ -43,11 +43,11 @@ import jakarta.persistence.UniqueConstraint;
 @Table(name = "ServiceOAuthNonce",
         uniqueConstraints = @UniqueConstraint(name = "uq_service_oauth_nonce",
                 columnNames = {"nonceKeyHash"}))
-public class ServiceOAuthNonce extends AbstractModel<Integer> {
+public class ServiceOAuthNonce extends AbstractModel<Long> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     // Fixed-length hash of the canonical (consumerKey, tokenId, nonce) tuple.
     // Uniqueness is enforced on this column so the unique index stays small and
@@ -67,11 +67,11 @@ public class ServiceOAuthNonce extends AbstractModel<Integer> {
     private Date dateCreated;
 
     @Override
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
